test: Update tests
