+++ api/src/test/java/com/fiap/hackathon/api/api/WorkingHourEntryApiTest.java
@@ -4,7 +4,6 @@ import com.fiap.hackathon.api.dto.workinghour.response.WorkingHourRegistryRespon
 import com.fiap.hackathon.controller.adapter.api.WorkingHourRegistryController;
 import com.fiap.hackathon.entity.WorkingHourRegistry;
 import com.fiap.hackathon.usecase.misc.secret.SecretUtils;
-import com.fiap.hackathon.usecase.misc.secret.TokenJwtSecret;
 import com.fiap.hackathon.api.misc.token.TokenJwtUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -12,8 +11,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 import java.time.Instant;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
@@ -38,12 +42,13 @@ class WorkingHourEntryApiTest {
     }
     @Test
     void shouldInsertWorkingHourEntry() {
-        var headerToken = "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJBdWRpZW5jZSIsImlzcyI6IlRFU1QtSVNTVUVSIiwiaWQiOiIxMjMiLCJuYW1lIjoiVGVzdGUiLCJlbWFpbCI6InRlc3RlQHRlc3RlLmNvbSIsInR5cGUiOiJVU0VSIiwiaWF0IjoxNzA2NDEzNDMyLCJqdGkiOiJmMDQyM2U0Yy03YzEwLTRiMzUtOGQ5NS1kYmZhZmMzY2ZkYTQifQ.mPm-fgLutJWpOCdCTAMJcS1XIHFclHD6kM3qOMEaJUw";
         var employeeId = "123";
+        var authorities = List.of(new SimpleGrantedAuthority("USER"));
         var registry = new WorkingHourRegistry(employeeId, Instant.now());
         var expected = WorkingHourRegistryResponseDto.toResponseDto(registry);
 
-        when(secretUtils.getTokenJwtSecret()).thenReturn(new TokenJwtSecret(TOKEN_SECRET, TOKEN_ISSUER));
+        Authentication authentication = new UsernamePasswordAuthenticationToken(employeeId, null, authorities);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
         when(controller.insert(any())).thenReturn(registry);
 
         WorkingHourRegistryResponseDto actual = api.insert();
